Moved dynamic rate limiter to seperate file
